refactor: introduce LegalPageLayout and CommerceTable components to standardize legal page rendering
